requirements.txt
Django
python-dotenv==1.1.0
psycopg2==2.9.10
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0

